scriptableObjectConnection

+++ Assets/Game/Scripts/ScriptMenu/SO/CharactersSO.cs
@@ -4,8 +4,8 @@ public class CharacterSO : ScriptableObject
 {
     public string characterName;
     public int health;
-    public int movementForce;
-    public int cooldownMovement;
+    public float movementForce;
+    public float cooldownMovement;
     public int baseAttackDamage;
     public float baseAttackCooldown;
     public float stunTime;
